+++ src/Efi.cs
@@ -93,6 +93,7 @@ public class Efi {
 			var pos = 6 + 2 * (Label.Length + 1);
 			var pathNodesEnd = pos + pathNodesLength;
 			DevicePathNodes = new List<DevicePathNode>();
+			Arguments = new byte[0];
 			while (pos + 4 <= pathNodesEnd) {
 				var len = BitConverter.ToUInt16(data, pos + 2);
 				if (len < 4 || pos + len > pathNodesEnd) {
@@ -378,9 +379,10 @@ public class Efi {
 	 *
 	 * @param label Label of the boot entry.
 	 * @param fileName File name of the boot entry.
+	 * @param alwaysCopyFromMS If true, do not preserve any existing data.
 	 * @param dryRun Don't actually create the entry.
 	 */
-	public static void MakeAndEnableBootEntry(string label, string fileName, bool dryRun = false) {
+	public static void MakeAndEnableBootEntry(string label, string fileName, bool alwaysCopyFromMS, bool dryRun = false) {
 		Variable msEntry = null, ownEntry = null;
 		UInt16 msNum = 0, ownNum = 0;
 
@@ -403,10 +405,11 @@ public class Efi {
 		Setup.Log($"Read EFI variable: {bootOrder}");
 		Setup.Log($"Read EFI variable: {bootCurrent}");
 		var bootOrderInts = new List<UInt16>(BytesToUInt16s(bootOrder.Data));
-		foreach (var num in BytesToUInt16s(bootCurrent.Data).Concat(bootOrderInts).Concat(Enumerable.Range(0, 0xffff).Select(i => (UInt16) i))) {
+		foreach (var num in BytesToUInt16s(bootCurrent.Data).Concat(bootOrderInts).Concat(Enumerable.Range(0, 0xff).Select(i => (UInt16) i))) {
 			var entry = GetVariable(String.Format("Boot{0:X04}", num));
 			if (entry.Data == null) {
-				if (ownEntry == null) {
+				// Use only Boot0000 .. Boot00FF because some firmwares expect that.
+				if (ownEntry == null && num < 0x100) {
 					ownNum = num;
 					ownEntry = entry;
 				}
@@ -434,12 +437,28 @@ public class Efi {
 			throw new Exception("MakeBootEntry: Windows Boot Manager not found.");
 		} else {
 			Setup.Log($"Read EFI variable: {msEntry}");
+			Setup.Log($"Read EFI variable: {ownEntry}");
 			// Make a new boot entry using the MS entry as a starting point.
 			var entryData = new BootEntryData(msEntry.Data);
-			entryData.Arguments = Encoding.UTF8.GetBytes(label + "\0");
+			if (!alwaysCopyFromMS && ownEntry.Data != null) {
+				entryData = new BootEntryData(ownEntry.Data);
+				// Shim expects the arguments to be a filename or nothing.
+				// But BCDEdit expects some Microsoft-specific data.
+				// Modify the entry so that BCDEdit still recognises it
+				// but the data becomes a valid UCS-2 / UTF-16LE file name.
+				var str = new string(entryData.Arguments.Take(12).Select(c => (char) c).ToArray());
+				if (str == "WINDOWS\0\x01\0\0\0") {
+					entryData.Arguments[8] = (byte) 'X';
+				} else if (str != "WINDOWS\0\x58\0\0\0") {
+					// Not recognized. Clear the arguments.
+					entryData.Arguments = new byte[0];
+				}
+			} else {
+				entryData.Arguments = new byte[0];
+				entryData.Label = label;
+				entryData.FileName = fileName;
+			}
 			entryData.Attributes = 1; // LOAD_OPTION_ACTIVE
-			entryData.Label = label;
-			entryData.FileName = fileName;
 			ownEntry.Attributes = 7; // EFI_VARIABLE_NON_VOLATILE | EFI_VARIABLE_BOOTSERVICE_ACCESS | EFI_VARIABLE_RUNTIME_ACCESS
 			ownEntry.Data = entryData.ToBytes();
 			SetVariable(ownEntry, dryRun);
@@ -471,7 +490,12 @@ public class Efi {
 			var bootOrderInts = new List<UInt16>(BytesToUInt16s(bootOrder.Data));
 			// Windows can't enumerate EFI variables, and trying them all is too slow.
 			// BootOrder + BootCurrent + the first 0xff entries should be enough.
-			foreach (var num in BytesToUInt16s(bootCurrent.Data).Concat(bootOrderInts).Concat(Enumerable.Range(0, 0xff).Select(i => (UInt16) i))) {
+			var seen = new HashSet<UInt16>();
+			foreach (var num in bootOrderInts.Concat(BytesToUInt16s(bootCurrent.Data)).Concat(Enumerable.Range(0, 0xff).Select(i => (UInt16) i))) {
+				if (seen.Contains(num)) {
+					continue;
+				}
+				seen.Add(num);
 				var entry = GetVariable(String.Format("Boot{0:X04}", num));
 				if (entry.Data != null) {
 					Setup.Log($"LogBootOrder: {entry}");
@@ -488,6 +512,9 @@ public class Efi {
 	public static string GetHackBGRTLog() {
 		try {
 			var log = GetVariable("HackBGRTLog", EFI_HACKBGRT_GUID);
+			if (log.Data == null) {
+				return "(null)";
+			}
 			return new string(BytesToUInt16s(log.Data).Select(i => (char)i).ToArray());
 		} catch (Exception e) {
 			return $"Log not found: {e.ToString()}";