+++ src/efi.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdint.h>
+#include <efi.h>
+#include <efilib.h>